# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CJ Vehicle Tracker API",
        "version": "1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1"
        }
    ],
    "paths": {
...
        "/mileage-trackers": {
            "get": {
                "tags": [
                    "Mileage Trackers"
                ],
                "summary": "Get mileage trackers",
                "operationId": "29749e30bf08318ce8ace2f917051df8",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MileageTrackerResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mileage trackers not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mileage Trackers"
                ],
                "summary": "Create a new mileage tracker",
                "operationId": "18deedd4927c6855f868da7149bb524d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMileageTrackerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mileage tracker created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MileageTrackerResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/mileage-trackers/get-paginated": {
            "get": {
                "tags": [
                    "Mileage Trackers"
                ],
                "summary": "Get paginated mileage trackers",
                "operationId": "0adf92e269ad684a9a9661b48fe03e8d",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 20
                        }
                    },
                    {
                        "name": "serialNo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "ss4er-sfsdf"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "available"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MileageTrackerPaginatedCollection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mileage trackers not found"
                    }
                }
            }
        },
        "/mileage-trackers/{id}": {
            "get": {
                "tags": [
                    "Mileage Trackers"
                ],
                "summary": "Get mileage tracker",
                "operationId": "defe251ad160903b376556ad67654be2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the mileage tracker to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MileageTrackerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mileage tracker not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mileage Trackers"
                ],
                "summary": "Update a mileage tracker",
                "operationId": "fe36bde11ccbd69086a3da1a66f357b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the mileage tracker to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMileageTrackerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mileage tracker updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MileageTrackerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mileage tracker not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mileage Trackers"
                ],
                "summary": "Delete a mileage tracker",
                "operationId": "22ab0867e5b1442bd2352a33e3ef9d13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the mileage tracker to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mileage tracker deleted successfully"
                    },
                    "404": {
                        "description": "Mileage tracker not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/mileage-trackers/{id}/is-deletable": {
            "get": {
                "tags": [
                    "Mileage Trackers"
                ],
                "summary": "Check mileage trackers delability",
                "operationId": "59da494e222c19628bfb6c8213f0a6ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the mileage trackers",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeletabilityResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mileage Tracker not found"
                    }
                }
            }
        },
        "/mileage-trackers/{id}/generate-key": {
            "put": {
                "tags": [
                    "Mileage Trackers"
                ],
                "summary": "Update a mileage tracker key",
                "operationId": "990b1cfdd5db6d2c6d5b56cb00016ae3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the mileage tracker to update key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mileage tracker key updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MileageTrackerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mileage tracker not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreMileageTrackerRequest": {
                "title": "MileageTracker Create Request",
                "required": [
                    "serialNo"
                ],
                "properties": {
                    "serialNo": {
                        "type": "string",
                        "example": "SE123123123"
                    }
                },
                "type": "object"
            },
            "UpdateMileageTrackerRequest": {
                "title": "MileageTracker Update Request",
                "required": [
                    "serialNo"
                ],
                "properties": {
                    "serialNo": {
                        "type": "string",
                        "example": "SE123123123"
                    }
                },
                "type": "object"
            },
...
            "DeletabilityResource": {
                "title": "Deletability Resource",
                "properties": {
                    "isDeletable": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "messages": {
                        "example": "[]"
                    }
                },
                "type": "object"
            },
...
            "MileageTrackerPaginatedCollection": {
                "title": "Mileage Tracker Paginated Collection",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MileageTrackerResource"
                        }
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64",
                        "example": 20
                    }
                },
                "type": "object"
            },
            "MileageTrackerResource": {
                "title": "Mileage Tracker Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "serialNo": {
                        "type": "string",
                        "example": "SE123123123"
                    },
                    "publicKey": {
                        "type": "string",
                        "example": "99882233-334455"
                    },
                    "privateKey": {
                        "type": "string",
                        "example": "7788-334433"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}